Rollback additional metric attributes
