+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -5,16 +5,8 @@
 * Added enrich functionality to metric instrumentation
   ([#1407](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1407)).
 
-  * New overload of `AddAspnetInstrumentation` now accepts a configuration delegate.
+  * New overload of `AddAspNetInstrumentation` now accepts a configuration delegate.
   * The `Enrich` can be used to add additional metric attributes.
-
-* Additional metric attributes will now be emitted for `http.server.duration`
-  ([#1407](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1407)):
-
-  * `net.host.name`
-  * `net.host.port`
-  * `http.flavor`
-  * `http.route`
 
 ## 1.6.0-beta.2
