+++ src/OpenTelemetry.Instrumentation.AspNet/Implementation/HttpInMetricsListener.cs
@@ -24,15 +24,12 @@ namespace OpenTelemetry.Instrumentation.AspNet.Implementation;
 
 internal sealed class HttpInMetricsListener : IDisposable
 {
-    internal const string HttpServerDurationMetricName = "http.server.duration";
-
-    private readonly HttpRequestRouteHelper routeHelper = new();
     private readonly Histogram<double> httpServerDuration;
     private readonly AspNetMetricsInstrumentationOptions options;
 
     public HttpInMetricsListener(Meter meter, AspNetMetricsInstrumentationOptions options)
     {
-        this.httpServerDuration = meter.CreateHistogram<double>(HttpServerDurationMetricName, "ms", "Measures the duration of inbound HTTP requests.");
+        this.httpServerDuration = meter.CreateHistogram<double>("http.server.duration", "ms", "Measures the duration of inbound HTTP requests.");
         TelemetryHttpModule.Options.OnRequestStoppedCallback += this.OnStopActivity;
         this.options = options;
     }
@@ -42,37 +39,16 @@ internal sealed class HttpInMetricsListener : IDisposable
         TelemetryHttpModule.Options.OnRequestStoppedCallback -= this.OnStopActivity;
     }
 
-    private static string GetHttpProtocolVersion(HttpRequest request)
-    {
-        var protocol = request.ServerVariables["SERVER_PROTOCOL"];
-        return protocol switch
-        {
-            "HTTP/1.1" => "1.1",
-            "HTTP/2" => "2",
-            "HTTP/3" => "3",
-            _ => protocol,
-        };
-    }
-
     private void OnStopActivity(Activity activity, HttpContext context)
     {
-        var request = context.Request;
-        var url = request.Url;
+        // TODO: This is just a minimal set of attributes. See the spec for additional attributes:
+        // https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/semantic_conventions/http-metrics.md#http-server
         var tags = new TagList
         {
-            { SemanticConventions.AttributeNetHostName, url.Host },
-            { SemanticConventions.AttributeNetHostPort, url.Port },
-            { SemanticConventions.AttributeHttpFlavor, GetHttpProtocolVersion(request) },
-            { SemanticConventions.AttributeHttpMethod, request.HttpMethod },
-            { SemanticConventions.AttributeHttpScheme, url.Scheme },
+            { SemanticConventions.AttributeHttpMethod, context.Request.HttpMethod },
+            { SemanticConventions.AttributeHttpScheme, context.Request.Url.Scheme },
             { SemanticConventions.AttributeHttpStatusCode, context.Response.StatusCode },
         };
-
-        var template = this.routeHelper.GetRouteTemplate(request);
-        if (!string.IsNullOrEmpty(template))
-        {
-            tags.Add(SemanticConventions.AttributeHttpRoute, template);
-        }
 
         if (this.options.Enrich is not null)
         {
