--- src/OpenTelemetry.Instrumentation.AspNet/Implementation/HttpRequestRouteHelper.cs
@@ -1,61 +1,0 @@
-// <copyright file="HttpRequestRouteHelper.cs" company="OpenTelemetry Authors">
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// </copyright>
-
-using System;
-using System.Web;
-using System.Web.Routing;
-
-namespace OpenTelemetry.Instrumentation.AspNet.Implementation;
-
-/// <summary>
-/// Helper class for processing http requests.
-/// </summary>
-internal class HttpRequestRouteHelper
-{
-    private readonly PropertyFetcher<object> routeFetcher = new("Route");
-    private readonly PropertyFetcher<string> routeTemplateFetcher = new("RouteTemplate");
-
-    /// <summary>
-    /// Extracts the route template from the <see cref="HttpRequest"/>.
-    /// </summary>
-    /// <param name="request">The <see cref="HttpRequest"/> being processed.</param>
-    /// <returns>The route template or <see langword="null"/>.</returns>
-    internal string? GetRouteTemplate(HttpRequest request)
-    {
-        var routeData = request.RequestContext.RouteData;
-
-        string? template = null;
-        if (routeData.Values.TryGetValue("MS_SubRoutes", out object msSubRoutes))
-        {
-            // WebAPI attribute routing flows here. Use reflection to not take a dependency on microsoft.aspnet.webapi.core\[version]\lib\[framework]\System.Web.Http.
-
-            if (msSubRoutes is Array attributeRouting && attributeRouting.Length == 1)
-            {
-                var subRouteData = attributeRouting.GetValue(0);
-
-                _ = this.routeFetcher.TryFetch(subRouteData, out var route);
-                _ = this.routeTemplateFetcher.TryFetch(route, out template);
-            }
-        }
-        else if (routeData.Route is Route route)
-        {
-            // MVC + WebAPI traditional routing & MVC attribute routing flow here.
-            template = route.Url;
-        }
-
-        return template;
-    }
-}
